+++ priv/repo/migrations/20231024100724_add_type_to_grant_requests.exs
@@ -1,6 +1,5 @@
 defmodule TdDd.Repo.Migrations.AddTypeToGrantRequests do
   use Ecto.Migration
-  alias Ecto.Multi
 
   @delete_data true