+++ priv/repo/migrations/20240608082444_add_units_index_linage.exs
@@ -1,0 +1,18 @@
+defmodule TdDd.Repo.Migrations.AddUnitsIndexLinage do
+  use Ecto.Migration
+
+  def up do
+    execute(
+      "CREATE INDEX units_nodes_node_id_hash_index ON public.units_nodes USING hash (node_id);"
+    )
+
+    execute(
+      "CREATE INDEX units_nodes_node_id_btree_index ON public.units_nodes USING btree (node_id);"
+    )
+  end
+
+  def down do
+    execute("DROP INDEX IF EXISTS units_nodes_node_id_btree_index;")
+    execute("DROP INDEX IF EXISTS units_nodes_node_id_hash_index;")
+  end
+end